+++ Controllers/Frontend/FatchipCTAjax.php
@@ -1,7 +1,5 @@
 <?php
 
-/* vim: set expandtab tabstop=4 shiftwidth=4 softtabstop=4: */
-
 /**
  * The Computop Shopware Plugin is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
@@ -115,7 +113,7 @@ class Shopware_Controllers_Frontend_FatchipCTAjax extends Enlight_Controller_Act
         $currency = $basket["sBasket"]["sCurrencyName"];
         $orderDesc = "Test";
 
-        $payment = $this->paymentService->getPaymentClass('AmazonPay', $this->config);
+        $payment = $this->paymentService->getPaymentClass('AmazonPay');
 
         $requestParams = $payment->getAmazonSCOParams(
             $this->session->offsetGet('fatchipCTPaymentPayID'),
@@ -152,7 +150,7 @@ class Shopware_Controllers_Frontend_FatchipCTAjax extends Enlight_Controller_Act
         $this->session->offsetSet('fatchipCTAmazonReferenceID', $referenceId);
         $orderDesc = "Test";
 
-        $payment = $this->paymentService->getPaymentClass('AmazonPay', $this->config);
+        $payment = $this->paymentService->getPaymentClass('AmazonPay');
 
         $requestParams = $payment->getAmazonGODParams(
             $this->session->offsetGet('fatchipCTPaymentPayID'),
@@ -200,7 +198,7 @@ class Shopware_Controllers_Frontend_FatchipCTAjax extends Enlight_Controller_Act
         $currency = 'EUR';
         $orderDesc = "Test";
 
-        $payment = $this->paymentService->getPaymentClass('AmazonPay', $this->config);
+        $payment = $this->paymentService->getPaymentClass('AmazonPay');
 
         $requestParams = $payment->getAmazonSODParams(
             $this->session->offsetGet('fatchipCTPaymentPayID'),