+++ Subscribers/Backend/OrderList.php
@@ -1,7 +1,4 @@
 <?php
-
-/* vim: set expandtab tabstop=4 shiftwidth=4 softtabstop=4: */
-
 /**
  * The Computop Shopware Plugin is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
@@ -30,12 +27,9 @@
 namespace Shopware\Plugins\FatchipCTPayment\Subscribers\Backend;
 
 use Enlight_Hook_HookArgs;
-
 use Enlight\Event\SubscriberInterface;
-
 use Shopware\Models\Order\Detail;
 use Shopware\Models\Attribute\OrderDetail as Attribute_OrderDetail;
-
 use RuntimeException;
 use Exception;
 
@@ -44,7 +38,7 @@ use Exception;
  *
  * @package Shopware\Plugins\FatchipCTPayment\Subscribers
  */
-class AfterBackendOrderGetListHook implements SubscriberInterface
+class OrderList implements SubscriberInterface
 {
     /**
      * return array with all subscribed events
@@ -115,11 +109,6 @@ class AfterBackendOrderGetListHook implements SubscriberInterface
             throw new RuntimeException('Unknown attribute base class');
         }
 
-        // roman.wedemeier@fatchip.de: 19.12.2018
-        //
-        // Before, here was an additional check, whether given Detail $detail
-        // has an Attribute_OrderDetail, but i think, that at this point it
-        // should be clear, that no OrderDetailAttribute exists on given Detail
         $attribute = new Attribute_OrderDetail();
 
         $detail->setAttribute($attribute);