+++ Subscribers/ControllerPath.php
@@ -1,6 +1,4 @@
 <?php
-/* vim: set expandtab tabstop=4 shiftwidth=4 softtabstop=4: */
-
 /**
  * The Computop Shopware Plugin is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
@@ -28,12 +26,10 @@
 
 namespace Shopware\Plugins\FatchipCTPayment\Subscribers;
 
-use Enlight\Event\SubscriberInterface;
-
 /**
  * provide paths to custom controllers
  */
-class ControllerPath implements SubscriberInterface
+class ControllerPath extends AbstractSubscriber
 {
 
     /**