+++ CHANGELOG.md
@@ -1,11 +1,11 @@
 # Changelog - Shopware Computop Payment Connector
 
 ## 1.0.28
-Pre-Released 2019-10-28
-* Rebranding: updated Computop icons and images
+Pre-Released 2019-12-10
+* Rebranding: updated Computop icons and images 
 * Refactored: unneccessary code removed
 * Fixed: attribute address flush / by @rdss-zmehmedovic
-* Added two new Klarna payment methods: Klarna SOFORT, Klarna Lastschrift
+* Added two new Klarna payment methods: Klarna SOFORT, Klarna Lastschrift 
 * Fixed: Amazon Pay address selection
 * Added: Configuration options for active credit card brands in silent mode